KD8ZRC.Flight.NBP3.History

Description

Functions to deal with coordinate history tracking.

Documentation

saveHistory

Arguments

:: FilePath
The file to save history to.

-> TelemetryLine
The downlink packet containing the latest data

-> IO ()

To store the history of a flight, we write out a simple JSON file. This is not part of the standard callback set for now, because different frontends and projects have different requirements for what needs to be saved.

writePktHistory :: FilePath -> ParsedPacketCallback TelemetryLine