KD8ZRC.Flight.NBP4.History

Description

Functions to deal with coordinate history tracking.

Documentation

newtype HistorySaveCallback t

Constructors

HistorySaveCallback (FilePath -> t -> IO ())

saveCoordinateHistory :: HistorySaveCallback TelemetryLine

To store the history of a flight, we write out a simple JSON file. This is not part of the standard callback set for now, because different frontends and projects have different requirements for what needs to be saved.

This saves only the coordinates, not other data in the packet. This is useful for sending a history of where the balloon has been to frontends so they can generate a visual path based on the numbers, without having to waste bandwidth sending historical data that will never be used.

saveFullHistory :: HistorySaveCallback TelemetryLine

However, there are cases where we *do* care about the other data, in a historical, machine-readable sense, such as when we want to generate graphs after a flight.

writePktHistory :: FilePath -> HistorySaveCallback t -> ParsedPacketCallback t

Take a filepath and one of the above callbacks and write the appropriate historical data to the file (depending on which callback above is used).

sendWSHistory :: FilePath -> WebsocketOnConnectCallback

Send data from the given JSON file to websocket clients as they connect.