KD8ZRC.Flight.NBP4.Types

Description

Types for NBP4

Documentation

type Meters = Double

data TelemetryLine

Constructors

TelemetryLine

Fields

_rawLine :: ByteString

We keep the raw line around for any callbacks that need it, but also so that we can define a HasCRC instance that makes use of it.

_callsign :: Text

_coordinates :: Coordinate

_altitude :: Meters

_time :: UTCTime

_voltage :: Double

_crc :: Integer

voltage :: Lens' TelemetryLine Double

time :: Lens' TelemetryLine UTCTime

rawLine :: Lens' TelemetryLine ByteString

crc :: Lens' TelemetryLine Integer

coordinates :: Lens' TelemetryLine Coordinate

callsign :: Lens' TelemetryLine Text

altitude :: Lens' TelemetryLine Meters

newtype JSONCoordinate

Constructors

JSONCoordinate Coordinate